Fixing bug: It wrote Right(...) to screen after a successful process
After a successful process it would write

      Right(Template applied in ./testingthis)

to screen. This was due to a `map` that should have been a `flatMap`
resulting in a Right[Nothing, String] where a String was expected.
Because the string was used in a string concatenation toString was
invoked on the Right[Nothing, String] instance.