Display a nicer error when we fail to find the maven metadata.
I added some logic to map a 404 response to a `Left` with an
appropriate error message, but this revealed that the left side of
this either was being discarded in fetchInfo where it assigned
`parameters`.

To fix that it I had to do a fair bit of refactoring, and dealt with
some whitespace abuse along the way. There's still a lot that's not
great about the code in g8.scala but at least it does now bubble up
the left side of the `Either` where it should.