Increase file/dir template pattern inclusivity
The template pattern used for matching against the underscore-formatted file/directory property substitutions only matched using the regex word character class. This excluded any special characters in the property names, as well as the dashes used in some of the long-form formatting options (such as "start-case", and "package-dir").

This changes the regex to use the custom character class [^$] (i.e. anything but a dollar sign) to allow a broader range of characters in both the property name as well as the formatter option name.