add stable dependency option to maven
This is an implementation of #164, which allows the user to specify
a "stable" dependency version using maven. "Stable", here implies that
the dependency is not a Release Candidate, Milestone Release or any
other non-final release. Basically it only allows releases that are
of the format: major.minor.patch.

This update does a couple of things:

1. Verifies that the latest version of a dependency is stable.
2. If the latest version of the dependency is not stable it looks through
all existing versions and finds the latest version that is stable.

The maven syntax has been extended to include an optional "stable" entry
which turns on a stable dependency:

```
maven(org, artifactId, stable)
```

I'm looking for some feedback as to whether this is an acceptable
approach to implement this solution. I'm happy to update it as necessary.