docs: add in Mill instructions
I recently made a Mill plugin for Giter8 so that Mill users can also
have the benefit of being able to test the generation of their
templates. I added some small notes and instructions on how to use it
under "Testing templates locally", but let me know if there is somewhere
else that is more fitting. Also if you'd like to not include external
plugins in the official docs, then no problem, feel free to close.