deps: exclude sshd-sftp
The reason for this pr is that the jgit version that is being used
brings in sshd-common (which is brought in by sshd-sftp) which has a
critical security vulnerability attached to it meaning GitHub complains.
You can see the advisory in https://github.com/advisories/GHSA-fhw8-8j55-vwgq.

I wasn't sure if I should pin or exclude, but I didn't see anything that
would use this and saw other exclusions, so I figured this was fine.

Note that another way to avoid this would be to bump jgit to the 6.x,
but I understand you don't want to do that as they drop support for Java
8.